+++ examples/src/examples.ts
@@ -31,6 +31,7 @@ import { MatterChains } from './06-physics/matter-chains';
 import { MatterCradle } from './06-physics/matter-cradle'; 
 import { MatterMixedShapes } from './06-physics/matter-mixed-shapes'; 
 import { ClothView as Cloth } from './06-physics/cloth'; 
+import { MatterBinding } from './06-physics/matter-binding';
 
 import { PixiHitTest } from './others/pixi-hittest';
 
@@ -68,6 +69,7 @@ import { PixiHitTest } from './others/pixi-hittest';
     MatterChains,
     MatterCradle,
     MatterMixedShapes,
+    MatterBinding,
     Cloth,
     PixiHitTest
 };